+++ packages/react/src/components/accordion/index.ts
@@ -1,2 +1,6 @@
 export { PlAccordion, PlAccordionItem } from './PlAccordion.js';
-export type { PlAccordionProps, PlAccordionItemProps } from './PlAccordion.js';
+export type {
+  PlAccordionHeadingLevel,
+  PlAccordionProps,
+  PlAccordionItemProps
+} from './PlAccordion.js';